requirements.txt
asgiref==3.3.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
